Add ebay client-go and application instance cache
This change is for caching ApplicationInstance which defined by tess.io API.
The cache will be used by k8s Service to query the ApplicationService which
it's mapped to.

Because this change only works with tess.io API, env variable
`PILOT_ENABLE_TESS_CUSTOM` is added for turning it off.