TESS-52683 fix DNS proxy doesn't resolve egress host to 240.240.x.x issue
When PILOT_FILTER_SIDECAR_CLUSTER_CONFIG is set to true, services could be
excluded for reducing CDS config size. This also affect NDS, and egress TCP
traffic could rely on DNS proxy to resolve host to a 240.240.x.x address.

This change defines `ndsServices` in SidecarScope for NDS, the existing
`services` is used for CDS as before.